Chapter 3 自然语言处理/CH 3_1 情感分析/requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tencentcloud-sdk-python==3.0.130
xlrd==1.2.0
